examples/pyproject.toml
[project]
name = "docutray-examples"
version = "0.1.0"
description = "Examples for the docutray SDK"
requires-python = ">=3.10"
dependencies = [
    "docutray",
    "python-dotenv>=1.2.1",
]

[tool.uv.sources]
docutray = { path = "..", editable = true }
